Introduce #searches interface for finders
Now #words and #target_word interfaces have gone and #searches is the
new one. #searches method should return a hash which includes one or
more pairs of a target name (string) and a list of names (array of objects
that respond to #to_s). e.g.:

  class TheFinder
    include DidYouMean::BaseFinder

    def searches
      {"nubmer" => %w(number character symbol)}
    end
  end

  TheFinder.new.suggestions # => ["number"]

This allows us to easily add more types of suggestions in one class:

  class MultipleTypeFinder
    include DidYouMean::BaseFinder

    def searches
      {
        "nubmer" => %w(number character symbol),
        "sybmol" => %w(number character symbol)
      }
    end
  end

  MultipleTypeFinder.new.suggestions # => ["number", "symbol"]