Use Levenshtein to reject undesired suggestions by Jaro-Nishijima-Winkler
This addresses #49. Jaro-Nishijima-Winkler distance still suggests too
many names. Let's use Levenshtein after the first selection and
specialcase if the word starts with the input so it can reject
undesired suggestions but still selects desired ones.

still not entirely happy about the implementation, but this is worth
giving a try.