Suggest private methods when a private method is called with args
Ruby will always raise an NoMethodError when a method is called with args.

  raiae
  # => NameError: undefined local variable or method `raiae' for main:Object

  raiae NoMethodError
  # => NoMethodError: undefined method `raiae' for main:Object

In the second example, did_you_mean should suggest the raise method.

Since this change, it'll look for the correct method name from private methods as well as public methods so that it can cover mroe cases.