Revert "Change the spell checker algorithm"
This reverts commit e2f5b2437f967565e4830eab6077f73ae166e0a7.

This change was made to special-case the NoMethodError caused by the `File.raed` call. But with the change 0a1b761 we can now properly suggest the `#read` method and we no longer need this change. Also, sometimes this algorithm made suggestions extremely inaccurate. For instance, `Minitest::Test` provides many methods starting with `assert_` and the spellcheker often suggested too many method names because of the nature of the Jaro-Winkler distance.