Make requires relative
Due to the nature of how ruby core is running the tests, the shipped ruby did_you_mean is higher in the path than this gem's lib directory. So when you require did_you_mean, you actually end up testing the shipped version instead of the source code. Because of that, this commit changes the requires to be relative so that we're guarunteed to be requiring the correct code. There doesn't really appear to be a standard way to do this within Ruby core itself, it looks like it's different depending on the gem.