Fix NoMethodError being dumpable with experimental
When experimental is turned on, the NoMethodError class was holding onto the @frame_binding instance variable, which made it unable to be dumped by Marshal. It doesn't appear that we actually need that instance variable once we've pulled out the instance varaibles names, so we can get rid of it.